Add linear regression through sum of deviations method.
Using two arrays, we find the mean of the two arrays. This mean is used to calculate the deviations on the variables. These deviations are used to find out Sum of Dev of independent var and sum of product of dev of dependent and independent vars. The use that to calc slope and intercept and then predict value of the dependent var.